Display error message if corresponding file can't be determined
When the corresponding file of a spec or target can't be determined,
display a useful error message to the user reflecting this, instead
of returning `nil` back to the caller (find-file*) which raises a
cryptic "(wrong-type-argument stringp nil)" error.

This implementation is conservative in that it doesn't change
`rspec-spec-or-target` behaviour unless a new, optional 'error
argument is passed in (added in the three call sites here), so that
any other programs that might call it won't have their behaviour
changed.